Add files via upload
This is a sample project that I created using Django in Python.  This is just about the framework design. The designs of the webpages has not been the main issue here and hence the webpages have just been left with a very basic HTML and CSS snippets, what has more value is the other files that have the main Python Code. In this project there is a home page, an admin page for access and a response submission page(Contact) that creates an interaction between the visitor and the admin, and all submissions to the form given are stored in the database which is accessed by the admin. Creating a basic framework layout was the main motive and hence the project has been kept simple and not very detailed-to-the-point.